change price and amount to BigDecimal

+++ backend/src/main/java/com/example/backend/entity/DishInCart.java
@@ -17,7 +17,7 @@ public class DishInCart {
     private int dishId;
     private String name;
     private String description;
-    private int amount;
-    private float onePiecePrice;
-    private float totalPrice;
+    private BigDecimal amount;
+    private BigDecimal onePiecePrice;
+    private BigDecimal totalPrice;
 }
